User details update
-aded a user details view (personal profile and foreign user profiles)
-supports different functionality for personal and non personal user
details
--the personal users profile contains all the users purchases with the
option to cancel them, all the user advertisements with a counter to the
ammount of buyers it has. The counter is clickable and displays a list
of all the buyers for the advert, the ammount they purchased and the
total price.  The names of the buyers are hyperlinks to their user
details.
--the non personal user profile displays information about the user and
a list of their advertisements.
-changed users names in the list ads view and details ads view to
hyperlinks to their profile